% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv.R
\docType{data}
\name{asv}
\alias{asv}
\title{Example ASV count table data}
\format{
## `asv`
A data frame with 32 rows and 13 columns:
\describe{
  \item{plot}{Some sample metadata such a plot in a field, etc}
  \item{tissue}{More sample metadata such as the tissue type}
  \item{sample}{A unique identifier for each sample (row)}
  \item{ASV1, ASV2, ..., ASV10}{Simulated ASV Counts}
}
}
\source{
Made from the example for \link{qc}
}
\usage{
asv
}
\description{
A very small toy asv table to use in examples.
}
\keyword{datasets}
