% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvDist.R
\name{asvDist}
\alias{asvDist}
\title{Function to make distance/dissimilarity data from an ASV table.}
\usage{
asvDist(
  asvTab,
  asvCols = NULL,
  method = "spearman",
  parallel = getOption("mc.cores", 1),
  clr_transform = T,
  edgeFilter = NULL,
  plot = F,
  outputname = NULL
)
}
\arguments{
\item{data}{}
}
\value{
A dataframe summarizing changepoint models for individual ASVs vs phenotypes.
}
\description{
I should inform this function using emd from pcvr.
That pipeline is probably really similar to what I want to have happen here.
In that case this should be ready to take either a dist object or a dataframe or a matrix similar to dist object
That way I can support all of vegdist.
}
\details{
So I need two functions here, asvDist and net
}
\examples{

# a<-qc(); b<-cal(a); c<-thresh(b); d<-asvDist(a)
print(load("/home/jsumner/Desktop/stargate/SINC/sincUtils/syncomBuilder/threshOutput.rdata"))
sp_dist<-asvDist(asv, method="spearman", clr_transform=T, edgeFilter=0.5)
unfiltered_sp_dist<-asvDist(asv, method="spearman", clr_transform=T, edgeFilter=NULL)
dim(sp_dist)
euc<-asvDist(asv, method="euclidean", clr_transform=F, edgeFilter=NULL)
bray<-asvDist(asv, method="bray", clr_transform=F, edgeFilter=NULL)
save(sp_dist, euc, bray, file="/home/jsumner/Desktop/stargate/SINC/sincUtils/syncomBuilder/asvDist_output.rdata")


}
\keyword{changepoint,}
\keyword{phenotype}
\keyword{regression,}
\keyword{threshold,}
