% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{asvNet}
\alias{asvNet}
\title{Function to perform network analysis on a distance/dissimilarity matrix of a microbial community.}
\usage{
asvNet(
  df,
  metadata = NULL,
  edge = NULL,
  thresh = NULL,
  metadata_join = "asv",
  thresh_join = "asv"
)
}
\arguments{
\item{df}{Long dataframe such as that returned from \link{asvDist}.}

\item{metadata}{Optional dataframe of metadata to attach to nodes. Metadata should belong to data at
the node level, so taxonomy for ASV nodes or genotype per sample nodes, but not genotype per ASV
since an ASV is present in many samples.}

\item{edge}{Column name of df to use for edge weighting. Typically this is the same as \code{method}
from \code{asvDist} used to make df.}

\item{thresh}{Optional output from \link{thresh}. If provided then P values of changepoints
will be added as node metadata.}

\item{metadata_join}{Column name of metadata that identifies nodes. Defaults to "asv".}

\item{thresh_join}{Column name of thresh that identifies nodes. Defaults to "asv" and there are no
cases currently where we expect it to be a good idea to change this.}
}
\value{
A named list with three elements:
\itemize{
   \item{"Nodes" is a dataframe of nodes and their metadata}
   \item{"Edges" is a dataframe of edges connecting nodes.}
   \item{"graph" is the igraph object used to generate the dataframes.}
}
}
\description{
Function to perform network analysis on a distance/dissimilarity matrix of a microbial community.
}
\examples{

taxa <- c("Bacteria", "Proteobacteria", "Betaproteobacteria", "Burkholderiales", 
          "Burkholderiaceae", "Paraburkholderia", NA)
taxa <- matrix(rep(taxa, 10), nrow = 10, byrow = TRUE)
colnames(taxa) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
rownames(taxa) <- paste0("ASV", 1:10)
# taxonomy data if used should have ASV names explicitly as a column
taxa_df <- as.data.frame(taxa)
taxa_df$asv <- rownames(taxa_df)

sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)
net_data <- asvNet(sp_dist, taxa_df, edge = "spearman")

}
\keyword{changepoint}
\keyword{network,}
