% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_da.R
\name{b_da}
\alias{b_da}
\title{Function to assess differential abundance of microbe counts or pseudo counts}
\usage{
b_da(
  df,
  col,
  predictors,
  zi_cutoff = 0.1,
  priors = NULL,
  intercept = TRUE,
  cores = getOption("mc.cores", 1),
  chains = 2,
  iter = 1000,
  ...
)
}
\arguments{
\item{df}{ASV table with metadata and counts}

\item{col}{ASV column to model}

\item{predictors}{Column names of predictors, can be discrete or continuous.}

\item{zi_cutoff}{Proportion of zeros to consider the distribution to be zero inflated,
defaults to 0.1.}

\item{priors}{Prior distributions as \code{brmsprior} objects, defaults are shown in details.
If left NULL (the default) then wide student priors will be used
(\code{brms::set_prior('student_t(3,0,5)',  class="b")}).}

\item{intercept}{Logical, should an intercept be used? Defaults to TRUE.}

\item{cores}{Argument to control number of cores to use in parallel \code{brms::brm()}.}

\item{chains}{Argument to control number of chains used by \code{brms::brm()}.}

\item{iter}{Argument to control number of iterations \code{brms::brm()}.}

\item{...}{Further arguments passed to \code{brms::brm()}.}
}
\value{
A brmsfit object
}
\description{
Function to assess differential abundance of microbe counts or pseudo counts
}
\details{
The prior distributions for a ZINB model are shown below, the only difference with
an NB model is that there is no zi parameter.

                  prior     class      coef group resp dpar nlpar lb ub       source
                  (flat)         b                                            default
                  (flat)         b levels                                (vectorized)
  student_t(3, 3.1, 2.5) Intercept                                            default
       gamma(0.01, 0.01)     shape                                  0         default
              beta(1, 1)        zi                                  0  1      default
}
\examples{
\dontrun{
if ("brms" \%in\% installed.packages()) {
  ex <- b_da(asv, col = "ASV1", predictors = "tissue", intercept = FALSE, chains = 1, iter = 500)
}
}
}
