% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{Function to demultiplex sequences either to single files or to multiple files for use with dada2.}
\usage{
demultiplex(
  fwd_reads,
  rev_reads,
  barcodes,
  fwd = "FWD",
  rev = "REV",
  name = c("Name", "Sample_name"),
  mode = c("dada", "general"),
  cores = getOption("mc.cores", 1),
  writeOut = TRUE,
  stat = FALSE
)
}
\arguments{
\item{fwd_reads}{A ShortReadQ object or a file path to a fastq file that will be read in with
\code{ShortRead::readFastq}.}

\item{rev_reads}{A ShortReadQ object or a file path to a fastq file that will be read in with
\code{ShortRead::readFastq}.}

\item{barcodes}{A dataframe of barcodes and metadata for naming or a file path to such a flat file
readable by \code{utils::read.delim}.}

\item{fwd}{The column name of the barcodes dataframe representing forward barcodes.
Defaults to "FWD".}

\item{rev}{The column name of the barcodes dataframe representing reverse barcodes.
Defaults to "REV".}

\item{name}{One or more column names from the barcodes dataframe to be used in naming samples.}

\item{mode}{Currently only "dada" is supported. This exists for tentative plans to have several
output modes.}

\item{cores}{Optionally the number of cores to run in parallel. This defaults to 1 if the "mc.cores"
option is not set.}

\item{writeOut}{Logical, should fastq/stats files be written out. Defaults to TRUE.
This can also be a path where data should be written to. Note that files are written as mode "w"
and this will throw an error if the file already exists.}

\item{stat}{Logical, should stats files be generated if writeOut is TRUE. Defaults to FALSE.}
}
\value{
Writes out fq files and summary statistics depending on writeOut and stat arguments.
If both are TRUE then summary stats are returned as a dataframe.
}
\description{
Function to demultiplex sequences either to single files or to multiple files for use with dada2.
}
\examples{
\dontrun{
r1 <- demultiplex(
  fwd_reads = "Plate11_R1_001.fastq.gz",
  rev_reads = "Plate11_R2_001.fastq.gz",
  barcodes = b1_formatted,
  name = c("well"),
  mode = "dada",
  cores = 10,
  writeOut = "test/",
  stat = TRUE
)
}
}
\keyword{dada2}
\keyword{demultiplex,}
