% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeFilter.R
\name{edgeFilter}
\alias{edgeFilter}
\title{Function to filter edges from networks generated by \code{asvNet}.}
\usage{
edgeFilter(net, filter, edge = "spearman")
}
\arguments{
\item{net}{Object returned from \link{asvNet}.}

\item{filter}{Value to filter edges for. If non-NULL then only edges with edgeWeight greater than
this value are kept. This can be a character vector or a numeric.}

\item{edge}{Optional weighting for edges. Must be present in the "edges" of net. Default of NULL
will show equal size edges between all connected nodes.}
}
\value{
A modified version of net with filtered edges (and nodes if any were now isolated).
}
\description{
Function to filter edges from networks generated by \code{asvNet}.
}
\examples{

# a<-qc(); b<-cal(a); c<-thresh(b); d<-asvDist(a) ; e<-net(d, thresh = c)
print(load("~/scripts/SINC/sincUtils/syncomBuilder/net_output.rdata"))
dim(net_data$edges)
net_data2 <- edgeFilter(net_data, 0.6)
dim(net_data2$edges)
net_data3 <- edgeFilter(net_data, "0.6")
dim(net_data3$edges)

}
