% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_thresh.R
\name{mcp_thresh}
\alias{mcp_thresh}
\alias{mcp_thresh.scbnet}
\title{Function to run bayesian changepoint models of individual microbes vs phenotypes or
communities/clusters vs phenotypes.}
\usage{
mcp_thresh(
  x,
  phenoCols,
  predCols = NULL,
  model = list(Y ~ 1, ~0 + X),
  cores = getOption("mc.cores", 1),
  calibratePheno = NULL,
  keep_models = FALSE,
  hypothesis = "X > 0",
  ...
)

\method{mcp_thresh}{scbnet}(
  x,
  phenoCols,
  predCols = NULL,
  model = list(Y ~ 1, ~0 + X),
  cores = getOption("mc.cores", 1),
  calibratePheno = NULL,
  keep_models = FALSE,
  hypothesis = "X > 0",
  asvTab = NULL,
  ...
)
}
\arguments{
\item{x}{An asv table or analogous dataframe with a row per observation and columns for traits.
Alternatively an \code{scbnet} object.}

\item{phenoCols}{A character vector of column names for phenotypes to be used in changepoint models.}

\item{predCols}{A character vector of columns representing microbes (predictor variables).
Defaults NULL where all column names containing the string "ASV" will be used.}

\item{model}{Changepoint model formula as used in mcp::mcp.
See [mcp documentation](https://lindeloev.github.io/mcp/articles/formulas.html) for examples.
If you are using multiple phenoCols or multiple predCols specify Y and X as outcome/predictor.}

\item{cores}{Number of cores to run in parallel.}

\item{calibratePheno}{An optional vector of column names to calibrate the phenotypes by.
This should generally correspond to those used in `cal` if the ASV table has been calibrated or
just represent confounders that you wish to remove effects from in the changepoint regression.
\code{lm} is used to get residuals of the phenotype after these effects are regressed out.}

\item{keep_models}{Should full model objects be kept? This can make thresh objects much larger.
Defaults to FALSE.}

\item{hypothesis}{A hypothesis to test, defaults to "X > 0" which uses the same logic as
the model formula to test phenoCols.}

\item{...}{Additional arguments passed to methods.}

\item{asvTab}{An ASV table used in making the \code{scbnet} object.}
}
\value{
A \code{thresh} object.
}
\description{
Function to run bayesian changepoint models of individual microbes vs phenotypes or
communities/clusters vs phenotypes.
}
\examples{

asv$biomass_z <- rnorm(nrow(asv))
tm <- mcp_thresh(asv, "biomass_z",
  model = list(
    Y ~ 1, # intercept,
    ~ 0 + X # joined slope after first changepoint
  )
)
tm

taxa <- c(
  "Bacteria", "Proteobacteria", "Betaproteobacteria", "Burkholderiales",
  "Burkholderiaceae", "Paraburkholderia", NA
)
taxa <- matrix(rep(taxa, 10), nrow = 10, byrow = TRUE)
colnames(taxa) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
rownames(taxa) <- paste0("ASV", 1:10)
# taxonomy data if used should have ASV names explicitly as a column
taxa_df <- as.data.frame(taxa)
taxa_df$asv <- rownames(taxa_df)
sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)
net_data <- asvNet(sp_dist, taxa_df, edge = "spearman_similarity")
net_data <- netClust(net = net_data, "components")
net_data <- netClust(net = net_data, pullNode(net_data,
  node = "ASV10",
  edge = "spearman_similarity", edgeFilter = 0.7, nodeCol = "asv"
))
net_data <- netClust(net = net_data, pullNode(net_data,
  node = "ASV1",
  edge = "spearman_similarity", edgeFilter = 0.7, nodeCol = "asv"
))
asv$biomass_z <- rnorm(nrow(asv))
tn <- mcp_thresh(net_data, phenoCols = "biomass_z", asvTab = asv)
tn

}
\keyword{changepoint,}
\keyword{phenotype}
\keyword{regression,}
\keyword{threshold,}
