% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeASVs.R
\name{mergeASVs}
\alias{mergeASVs}
\title{Function combine ASV tables that have been aligned with blastASVs}
\usage{
mergeASVs(
  blast,
  asv,
  taxa,
  shared_meta,
  check_taxa = TRUE,
  min_pident = 99.5,
  cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{blast}{A dataframe as returned by \link{blastASVs}.}

\item{asv}{A list of two ASV tables.}

\item{taxa}{Optional taxonomy information in the form returned by dada2 for both experiments.
Defaults to NULL.}

\item{shared_meta}{Optional column names present in both ASV tables that you wish to include in the
new ASV table output.}

\item{check_taxa}{Logical, should taxa be checked for overlap and the results included in the link
part of the output? Note even if this is false you will be able to tell if the taxa are misaligned
in the taxa part of the output if taxa is non NULL,}

\item{min_pident}{Minimum pident match returned by blastASVs to merge. Defaults to 99.}

\item{cores}{Optionally a number of cores to run in parallel using parallel::mclapply. Defaults to 1
if the "mc.cores" option is unset}
}
\value{
A named list of dataframes including "asv" and "links", optionally with "taxa"
}
\description{
Function combine ASV tables that have been aligned with blastASVs
}
\examples{

res <- read.csv("blast_2022_vs_2021.csv")
print(load("~/scripts/SINC/field_2021/microbiome/sinc_field_2021_asvTable.rdata"))
asv21 <- asv ; taxa21 <- taxa
colnames(asv21)[c(1,7)] <- c("genotype", "tissue")
print(load("~/scripts/SINC/field_2022/microbiome/2022_asvTable.rdata"))

blast <- res
asv <- list("d1" = asv22, "d2" = asv21)
taxa <- list("d1" = taxa22, "d2" = taxa21)
shared_meta <- c("genotype", "tissue") # shared metadata to keep from both tables

ex <- mergeASVs(blast, asv, taxa, shared_meta)

}
\keyword{ASVs}
\keyword{combine}
