% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netPlot.R
\name{net.plot}
\alias{net.plot}
\title{Function to plot networks generated from \code{asvNet} with various emphases.}
\usage{
net.plot(
  net,
  fill = NULL,
  shape = NULL,
  size = 3,
  edgeWeight = NULL,
  edgeFilter = NULL,
  thresh_below = 0.05,
  facet = NULL,
  method = "ggplot"
)
}
\arguments{
\item{net}{Object returned from \link{asvNet}.}

\item{fill}{Optional column name to fill points by. Accepts single column names, vectors of p-value
columns, or "thresh" which will match all p-value columns if the network was fit with threshold model
data from \link{thresh}.}

\item{shape}{Optional column name to use for node shapes. Accepts the same options as fill.}

\item{size}{Size for points. Passed to ggplot2::geom_point.}

\item{edgeWeight}{Optional weighting for edges. Must be present in the "edges" of net. Default of
NULL will show equal size edges between all connected nodes.}

\item{edgeFilter}{Optional value to filter edges for. If non-NULL then only edges with edgeWeight
greater than this value are kept.
This can be a character vector or a numeric.
Character vectors are interpreted as quantiles ("0.5" corresponds to the top 50 percent are kept).}

\item{thresh_below}{Significant cutoff if p-value columns are used for fill or shape.
Defaults to 0.05.}

\item{facet}{Optionally a variable to facet the plot on. This is meant to be used to separate
multiple networks created from \link{netcomi2scb}
in which case "netNumber" should be used.}

\item{method}{Method for visualization, defaults to "ggplot" but also accepts "plotly".}
}
\value{
A plot of the kind determined by the method argument.
}
\description{
Function to plot networks generated from \code{asvNet} with various emphases.
}
\examples{

taxa <- c("Bacteria", "Proteobacteria", "Betaproteobacteria", "Burkholderiales", 
          "Burkholderiaceae", "Paraburkholderia", NA)
taxa <- matrix(rep(taxa, 10), nrow = 10, byrow = TRUE)
colnames(taxa) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
rownames(taxa) <- paste0("ASV", 1:10)
# taxonomy data if used should have ASV names explicitly as a column
taxa_df <- as.data.frame(taxa)
taxa_df$asv <- rownames(taxa_df)

sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)
net_data <- asvNet(sp_dist, taxa_df, edge = "spearman")

net.plot(net_data, size = 3, edgeWeight = "spearman", thresh_below = 0.05)

}
