% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netClust.R
\name{netClust}
\alias{netClust}
\title{Function to cluster networks generated from \code{asvNet}.}
\usage{
netClust(net, method = "components", ...)
}
\arguments{
\item{net}{Object returned from \link{asvNet}}

\item{method}{Method to use for clustering. This can be a method from "components", "dbscan",
and "kmeans" or the output from pullNode, in which case that node and it's connections are labelled
as a cluster.}

\item{...}{Additional arguments passed to function picked by method.}
}
\value{
A named list (same as net) with three elements, same as \link{asvNet}:
\itemize{
   \item{"Nodes" is a dataframe of nodes and their metadata}
   \item{"Edges" is a dataframe of edges connecting nodes.}
   \item{"graph" is the igraph object used to generate the dataframes.}
}
}
\description{
Function to cluster networks generated from \code{asvNet}.
}
\details{
Each method will use a different function to cluster data according to the layout
         in the graph. Note that layouts in the graph are determined by \code{igraph::layout_nicely}.
\itemize{
   \item{"component" uses \code{igraph::components} to cluster data and
         requires no additional arguments.}
   \item{"dbscan" uses \code{dbscan::dbscan} to cluster data.
         This requires at least that the eps argument is set. See \code{?dbscan::dbscan}.}
   \item{"kmeans" uses \code{stats::kmeans} to cluster data.
         This requires at least that the centers argument is set.}
}
}
\examples{

taxa <- c("Bacteria", "Proteobacteria", "Betaproteobacteria", "Burkholderiales", 
          "Burkholderiaceae", "Paraburkholderia", NA)
taxa <- matrix(rep(taxa, 10), nrow = 10, byrow = TRUE)
colnames(taxa) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
rownames(taxa) <- paste0("ASV", 1:10)
# taxonomy data if used should have ASV names explicitly as a column
taxa_df <- as.data.frame(taxa)
taxa_df$asv <- rownames(taxa_df)
sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)
net_data <- asvNet(sp_dist, taxa_df, edge = "spearman")
net_data <- netClust(net = net_data, "components")

}
\keyword{changepoint}
\keyword{network,}
