% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcomi2scb.R
\name{netcomi2scb}
\alias{netcomi2scb}
\title{Function to translate NetCoMi objects into dataframes usable by other syncombuildr functions.}
\usage{
netcomi2scb(
  microNetObj,
  microNetProps = NULL,
  microNetComp = NULL,
  cutoff = "0.9",
  combine = TRUE
)
}
\arguments{
\item{microNetObj}{A microNetObj object as returned by \code{NetCoMi::netConstruct}}

\item{microNetProps}{A microNetProps object as returned by \code{NetCoMi::netAnalyze}}

\item{microNetComp}{A microNetComp object as returned by \code{NetCoMi::netCompare}}

\item{cutoff}{Optional value to filter edges for. If non-NULL then only edges with edgeWeight
greater than this value are kept.
This can be a character vector or a numeric.
Character vectors are interpreted as quantiles ("0.5" corresponds to the top 50% are kept).
Defaults to "0.9" .}

\item{combine}{Logical, should multiple networks have X/Y coordinates for nodes standardized to
make comparable plots later on? Defaults to TRUE.}
}
\value{
If the microNetObj contains one network then a named list with three elements is returned:
\itemize{
   \item{"Nodes" is a dataframe of nodes and their metadata}
   \item{"Edges" is a dataframe of edges connecting nodes.}
   \item{"graph" is the igraph object used to generate the dataframes.}
}
If the microNetObj contains two networks then a list of such lists is returned.
}
\description{
The main purpose of this function is to bring NetCoMi's metrics and testing into a format
that allows for more customized visualization such as with \link{net.plot}
}
\examples{

library(NetCoMi)
microNetObj <- NetCoMi::netConstruct(as.matrix(asv[, grepl("ASV", colnames(asv))]),
  measure = "spearman", sparsMethod = "t-test", alpha = 0.7
)
net <- netcomi2scb(microNetObj)

}
\keyword{changepoint}
\keyword{network,}
