% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadf.R
\name{pcadf}
\alias{pcadf}
\title{Function to run a PCA, plot and return the data with PC coordinates}
\usage{
pcadf(
  df = NULL,
  cols = NULL,
  color = NULL,
  returnData = TRUE,
  ncp = NULL,
  umap = FALSE,
  distance = "euclidean"
)
}
\arguments{
\item{df}{Dataframe to ordinate}

\item{cols}{columns to reduce dimensions of. Can be specified with names or positions. Defaults to
all column names containing "ASV".}

\item{color}{column name used to color points in the pca plot.}

\item{returnData}{Logical, should data be returned? Defaults to TRUE where data and a ggplot are
eturned.}

\item{ncp}{Optional, number of principal components to return attached to dataframe if data is
returned. Defaults to all.}

\item{umap}{Logical, should a UMAP also be performed? Defaults to FALSE. If TRUE then UAMP1 and 2
columns will be added and a list of 2 ggplots will be returned.}

\item{distance}{Distance to use. The default ("euclidean") will use \code{FactoMineR::PCA}, distances
supported by \code{vegan::vegdist} will use \code{ape::pcoa}.}
}
\value{
a ggplot, optionally a dataframe and FactoMineR::PCA output if returnData is TRUE.
}
\description{
This is broadly useful for quality control, especially when used with \link{taxaAg} in case the
ASV level trend does not seem to make sense.
}
\examples{

x <- pcadf(
  df = asv, cols = NULL,
  color = "tissue", returnData = TRUE, ncp = NULL
)
x$plot

}
\keyword{pca}
\keyword{umap}
