% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbnet_class.R
\name{print.scbnet}
\alias{print.scbnet}
\title{Print an \code{scbnet} object.}
\usage{
\method{print}{scbnet}(x, ...)
}
\arguments{
\item{x}{An object of class \code{scbnet}.}

\item{...}{further arguments, passed to print.default.}
}
\description{
Print an \code{scbnet} object.
}
\seealso{
\code{\link{summary.scbnet}}
}
