% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tresh_class.R
\name{print.thresh}
\alias{print.thresh}
\title{Print an \code{thresh} object.}
\usage{
\method{print}{thresh}(x, ...)
}
\arguments{
\item{x}{An object of class \code{thresh}.}

\item{...}{further arguments, passed to print.default.}
}
\description{
Print an \code{thresh} object.
}
\seealso{
\code{\link{summary.thresh}}
}
