% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullNode.R
\name{pullNode}
\alias{pullNode}
\title{Function to highlight subsets of networks generated from \code{asvNet} or \link{netcomi2scb}.}
\usage{
pullNode(
  net,
  node,
  edge = NULL,
  edgeFilter = NULL,
  plot = TRUE,
  nodeCol = "asv",
  keepNames = FALSE
)
}
\arguments{
\item{net}{Network list returned from \link{asvNet} or \link{netcomi2scb}.}

\item{node}{A node name (or vector of names) to extract connections to.}

\item{edge}{Optional weighting for edges. Must be present in the "edges" of net. Default of NULL
will show equal size edges between all connected nodes.}

\item{edgeFilter}{Optional value to filter edges for. If non-NULL then only edges with edgeWeight
greater than this value are kept.
This can be a character vector or a numeric.
Character vectors are interpreted as quantiles ("0.5" corresponds to the top 50 percent are kept).}

\item{plot}{Logical, should data be plotted using \link{net.plot}.}

\item{nodeCol}{Column of node information to find the node in. Defaults to "asv".}

\item{keepNames}{Logical, should other nodes be filled in the plot as their node name (TRUE) or all
as "other" (FALSE)?
Defaults to FALSE.}
}
\value{
A list with a subset network and optionally a ggplot.
}
\description{
Function to highlight subsets of networks generated from \code{asvNet} or \link{netcomi2scb}.
}
\examples{

taxa <- c(
  "Bacteria", "Proteobacteria", "Betaproteobacteria", "Burkholderiales",
  "Burkholderiaceae", "Paraburkholderia", NA
)
taxa <- matrix(rep(taxa, 10), nrow = 10, byrow = TRUE)
colnames(taxa) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
rownames(taxa) <- paste0("ASV", 1:10)
# taxonomy data if used should have ASV names explicitly as a column
taxa_df <- as.data.frame(taxa)
taxa_df$asv <- rownames(taxa_df)
sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)
net_data <- asvNet(sp_dist, taxa_df, edge = "spearman")

sub_net <- pullNode(net_data, node = "ASV10", edge = "spearman", plot = TRUE, nodeCol = "asv")

}
