% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_class.R
\name{[.thresh}
\alias{[.thresh}
\alias{`[.thresh`}
\alias{Thresh}
\alias{objects}
\alias{have}
\alias{variable}
\alias{lengths}
\alias{and}
\alias{it's}
\alias{common}
\alias{to}
\alias{want}
\alias{subset}
\alias{them}
\alias{based}
\alias{on}
\alias{phenotypes}
\alias{or}
\alias{predictors.}
\alias{Specialized}
\alias{bracket}
\alias{methods}
\alias{are}
\alias{implemented}
\alias{support}
\alias{that.}
\title{Bracket subsetting for thresh objects}
\usage{
\method{[}{thresh}(x, i)
}
\arguments{
\item{x}{thresh object}

\item{i}{indexing, logical and numeric vectors are handled specially to subset the thresh object.
Other options are handled per a normal list}
}
\description{
Bracket subsetting for thresh objects
}
\examples{
asv$biomass_z <- rnorm(nrow(asv))
tm <- thresh(asv, "biomass_z")
tm["predictor"] # pulls list
tm[1:2] # subsets lists to make a smaller thresh object
}
