% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbnet_class.R
\name{summary.scbnet}
\alias{summary.scbnet}
\title{Summarize an \code{scbnet} object.}
\usage{
\method{summary}{scbnet}(object, ...)
}
\arguments{
\item{object}{An object of class \code{scbnet}.}

\item{...}{further arguments, passed to summary.default}
}
\description{
Summarize an \code{scbnet} object.
}
