% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tresh_class.R
\name{summary.thresh}
\alias{summary.thresh}
\title{Summarize an \code{thresh} object.}
\usage{
\method{summary}{thresh}(object, ...)
}
\arguments{
\item{object}{An object of class \code{thresh}.}

\item{...}{further arguments, passed to summary.default}
}
\description{
Summarize an \code{thresh} object.
}
