% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv.R
\docType{data}
\name{taxa}
\alias{taxa}
\title{Example ASV count table data}
\format{
## `taxa`
A matrix with 10 rows and 6 columns, corresponding to the asv data:
\describe{
  \item{Kingdom}{Simulated kingdom data}
  \item{Phylum}{Simulated phylum data}
  \item{Class}{Simulated class data}
  \item{Order}{Simulated order data}
  \item{Family}{Simulated family data}
  \item{Genus}{Simulated genus data}
}
}
\source{
Pulled from a SINC dataset
}
\usage{
taxa
}
\description{
A set of taxa information for use in examples.
}
\keyword{datasets}
