% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaAg.R
\name{taxaAg}
\alias{taxaAg}
\title{Aggregate an ASV table to some taxonomic level}
\usage{
taxaAg(x = NULL, tx = NULL, taxaLevel = "Class")
}
\arguments{
\item{x}{ASV table, asv columns thought start with "ASV" prefix.}

\item{tx}{Taxonomy table in dada2 format. Rownames should correspond to ASV names in the same format as x's column names.}

\item{taxaLevel}{A level of the taxonomy in `tx` to aggregate data to. Must match a colname of `tx`.}
}
\description{
Aggregate an ASV table to some taxonomic level
}
\keyword{dirichlet}
\keyword{network}
