% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaAg.R
\name{taxaSums}
\alias{taxaSums}
\title{Make named vector of taxonomy sums}
\usage{
taxaSums(x = NULL, tx = NULL, taxaLevel = "Class")
}
\arguments{
\item{x}{ASV table, asv columns thought start with "ASV" prefix.}

\item{tx}{Taxonomy table in dada2 format. Rownames should correspond to ASV names in the same
format as x's column names. If this is a matrix it will be converted to a data.frame.}

\item{taxaLevel}{A level of the taxonomy in `tx` to aggregate data to. Must match a colname of `tx`.}
}
\description{
Make named vector of taxonomy sums
}
\examples{
taxaSums(asv, taxa, "Class")

}
\keyword{dirichlet}
\keyword{taxonomy,}
