% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tresh_class.R
\docType{class}
\name{thresh-class}
\alias{thresh-class}
\title{Class \code{thresh} for output from the \code{syncomBuildR::asvNet} function.}
\description{
Class for networks as used in syncomBuildR.
}
\details{
See \code{methods(class = "thresh")} for an overview of available methods.
}
\section{Slots}{

\describe{
\item{\code{nodes}}{Data frame of node information}

\item{\code{edges}}{Data frame of edge information}

\item{\code{graph}}{igraph object}

\item{\code{control}}{Additional (optional) information about the network}
}}

\seealso{
\code{\link{asvNet}}
}
