% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.R
\name{thresh}
\alias{thresh}
\title{Function to run changepoint models of individual microbes vs phenotypes or
communities/clusters vs phenotypes.}
\usage{
thresh(
  asvTab,
  phenoCols,
  asvCols = NULL,
  model = "hinge",
  cores = getOption("mc.cores", 1),
  calibratePheno = NULL,
  p.adjust.method = "none"
)
}
\arguments{
\item{asvTab}{An asv table or analogous dataframe with a row per observation and columns for traits.}

\item{phenoCols}{A character vector of column names for phenotypes to be used in changepoint models.}

\item{asvCols}{A character vector of columns representing microbes (predictor variables).
Defaults NULL where all column names containing the string "ASV" will be used.}

\item{model}{Type of changepoint model in chngpt::chngptm labeling convention. Currently hinge,
upperhinge, and segmented are supported. See Figure 2.1 of the chngpt
[vignette](https://cran.r-project.org/web/packages/chngpt/vignettes/chngpt-vignette.pdf).}

\item{cores}{Number of cores to run in parallel.}

\item{calibratePheno}{An optional vector of column names to calibrate the phenotypes by.
This should generally correspond to those used in `cal` if the ASV table has been calibrated or
just represent confounders that you wish to remove effects from in the changepoint regression.}

\item{p.adjust.method}{A method to adjust for multiple testing, defaults to "none". Available options
are shown with \code{stats::p.adjust.methods}.
\code{lm} is used to get residuals of the phenotype after these effects are regressed out.}
}
\value{
A \code{thresh} object.
}
\description{
Function to run changepoint models of individual microbes vs phenotypes or
communities/clusters vs phenotypes.
}
\examples{

asv$biomass_z <- rnorm(nrow(asv))
tm <- thresh(asv, "biomass_z")
tm

}
\keyword{changepoint,}
\keyword{phenotype}
\keyword{regression,}
\keyword{threshold,}
