% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshUpset.R
\name{threshUpset}
\alias{threshUpset}
\title{Function to make upset plots of thresh output}
\usage{
threshUpset(thresh, cores = getOption("mc.cores", 1), method = "fdr")
}
\arguments{
\item{thresh}{output from the thresh function}

\item{cores}{Number of cores to use in summarizing data. This defaults to 1 if the "mc.cores" option
is not set.}

\item{method}{The p-adjustment method to be used. Can be any suffix from ?p.adjust.methods}
}
\value{
A list of ggplots
}
\description{
Function to make upset plots of thresh output
}
\examples{

# a<-qc(); b<-cal(a); c<-thresh(b)
print(load("/home/jsumner/Desktop/stargate/SINC/sincUtils/syncomBuilder/cal_output.rdata"))
asv <- are_c[[1]]
zinbCalibrated <- are_c[[2]][are_c[[2]]$model == "ZINB", "asv"]
asv <- are_c[[1]][, c("tissue", "plot", "row", "genotype", "biomass", "sd", zinbCalibrated)]

threshMods <- thresh(
  asvTab = asv, phenoCols = "biomass",
  model = "hinge", cores = 10, calibratePheno = "genotype"
)
threshUpset(threshMods, 1)

}
\keyword{changepoint,}
\keyword{phenotype}
\keyword{regression,}
\keyword{threshold,}
