% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvDist.R
\name{asvDist}
\alias{asvDist}
\title{Function to make distance/dissimilarity data from an ASV table.}
\usage{
asvDist(
  asvTab,
  asvCols = NULL,
  method = "spearman",
  parallel = getOption("mc.cores", 1),
  clr_transform = TRUE,
  edgeFilter = NULL,
  plot = FALSE,
  returnASVs = TRUE
)
}
\arguments{
\item{asvTab}{ASV table to use. Should have rows as samples and columns representing ASVs.}

\item{asvCols}{Either column names, positions, or a single character string to use as regex to
find all asv columns.}

\item{method}{A distance method compatible with \code{vegan::vegdist}, a correlation
coefficient for use with \code{Hmisc::rcorr} (spearman or pearson), or one of "blomqvist" or
"gaussrank", which are implemented here in R and therefore will be slower.}

\item{parallel}{How many cores to use in parallel. The parallelized component only comes into play
with correlation coefficient methods and is not particularly heavy, so more cores will have a
minimal speed improvement.}

\item{clr_transform}{Should ASV counts be transformed with a centered log ratio?
This defaults to TRUE to alleviate the compositional nature of the data.}

\item{edgeFilter}{A filter for the edges returned. This defaults to NULL where no filtering is done.
Optionally this can be numeric, in which case distances/correlations below that number are removed,
or it can be a number as a string in which case that quantile is used
(ie, "0.5" filters for above the median.).}

\item{plot}{Logical, should the data be plotted? Defaults to FALSE.}

\item{returnASVs}{Should ASVs or samples be considered the experimental unit?
Defaults to TRUE in which case distances/correlations between ASVs across samples are returned.
If FALSE then samples are compared by their ASV composition. Note that different methods are
appropriate depending on whether samples or ASVs are being considered "nodes".}
}
\value{
A dataframe showing pairwise correlations between individual ASVs/samples. If plot is TRUE
then a list is returned with the dataframe as the "data" element and the plot as the "plot" element.
}
\description{
Function to make distance/dissimilarity data from an ASV table.
}
\details{
If "spearman" or "pearson" methods are used then euclidean distance is also calculated from those
correlations using the cosine theorem. That distance is bounded on [0, 2] corresponding to
correlations of -1 and 1, respectively. See examples.
}
\examples{

sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)

# note how distance is calculated:
x <- seq(-1, 1, 0.0001) # correlation range
y <- sqrt(2 * (1 - x)) # cosine transformation to euclidean distance
plot(x, y, type = "l")

}
\keyword{asv}
