% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastASVs.R
\name{blastASVs}
\alias{blastASVs}
\title{Function to use BLAST to align ASV sequences across multiple datasets.}
\usage{
blastASVs(
  seqs1,
  seqs2,
  cutoff = 97.5,
  db = NULL,
  maxMatches = Inf,
  cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{seqs1}{A named list of ASV sequences. Generally the names should be ASV numbers.}

\item{seqs2}{An optional second named list of ASV sequences. If db is left NULL then this is used
to make a blast database and seqs1 will be blasted against it before it is removed.}

\item{cutoff}{A percentage identity required to return a match. Defaults to 97.5}

\item{db}{An optional database to use. Note if this is supplied then seqs2 is ignored.}

\item{maxMatches}{An optional number of maximum matches to return per sequence in seqs1.
Defaults to Inf, which will return all matches above the cutoff.}

\item{cores}{Number of cores optionally to run in parallel. Defaults to 1 if "mc.cores" is not set.}
}
\value{
A dataframe of blast results
}
\description{
Note that currently no attempt is made to resolve dependencies for this function so you will need to
have rBLAST from github (\code{devtools::install_github("mhahsler/rBLAST")}) and Biostrings/seqinr
already installed. I'll get to at some point.
}
\examples{

if ("rBLAST" \%in\% installed.packages() && rBLAST::has_blast()) {
  seqs1 <- list(
    ASV9 = paste0(
      "GTGCCAGCAGCCGCGGTAATACGAAGGGGGCTAGCGTTGCTCGGAATCACTGGGCGT",
      "AAAGGGTGCGTAGGCGGGTCTTTAAGTCAGGGGTGAAATCCTGGAGCTCAACTCCAGAACTGCCT",
      "TTGATACTGAAGATCTTGAGTTCGGGAGAGGTGAGTGGAACTGCGAGTGTAGAGGTGAAATTCGTA",
      "GATATTCGCAAGAACACCAGTGGCGAAGGCGGCTCACTGGCCCGATACTGACGCTGAGGCACGAAAGCGT",
      "GGGGAGCAAACAGGATTAGATACCCTGGTAGTCCACGCCGTAAACGATGAATGCCAGCCGTTAGTGGGTT",
      "TACTCACTAGTGGCGCAGCTAACGCTTTAAGCATTCCGCCTGGGGAGTACGGTCGCAAGATTAAAACTCAAATGAATTGACGG"
    ),
    ASV10 = paste0(
      "GTGCCAGCCGCCGCGGTAATACGAAGGGGGCTAGCGTTGCTCGGAATCACTGGGCGTAAAGGGTGCGTAGGCGGGTCT",
      "TTAAGTCAGGGGTGAAATCCTGGAGCTCAACTCCAGAACTGCCTTTGATACTGAAGATCTTGAGTTCGGGAGAGGTGAGT",
      "GGAACTGCGAGTGTAGAGGTGAAATTCGTAGATATTCGCAAGAACACCAGTGGCGAAGGCGGCTCACTGGCCCGATACT",
      "GACGCTGAGGCACGAAAGCGTGGGGAGCAAACAGGATTAGATACCCTGGTAGTCCACGCCGTAAACGATGAATGCCAGCC",
      "GTTAGTGGGTTTACTCACTAGTGGCGCAGCTAACGCTTTAAGCATTCCGCCTGGGGAGTACGGTCGCAAGATTAAAACT",
      "CAAATGAATTGACGG"
    )
  )
  seqs2 <- list(
    ASV1 = paste0(
      "GTGCCAGCAGCCGCGGTAATACAGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCGCGTAGGTGGTTTGTTAAGT",
      "TGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCAAAACTGACAAGCTAGAGTATGGTAGAGGGTGGTGGAATTTCCT",
      "GTGTAGCGGTGAAATGCGTAGATATAGGAAGGAACACCAGTGGCGAAGGCGACCACCTGGACTGATACTGACACTGAGGTGC",
      "GAAAGCGTGGGGAGCAAACAGGATTAGATACCCTGGTAGTCCACGCCGTAAACGATGTCAACTAGCCGTTGGGAGCCTTG",
      "AGCTCTTAGTGGCGCAGCTAACGCATTAAGTTGACCGCCTGGGGAGTACGGCCGCAAGGTTAAAACTCAAATGAATTGACGG"
    ),
    ASV2 = paste0(
      "GTGCCAGCCGCCGCGGTAATACAGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCGCGTAGGTGGTTTGTTAA",
      "GTTGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCAAAACTGACAAGCTAGAGTATGGTAGAGGGTGGTGGAATT",
      "TCCTGTGTAGCGGTGAAATGCGTAGATATAGGAAGGAACACCAGTGGCGAAGGCGACCACCTGGACTGATACTGACACTGA",
      "GGTGCGAAAGCGTGGGGAGCAAACAGGATTAGATACCCTGGTAGTCCACGCCGTAAACGATGTCAACTAGCCGTTGGGAG",
      "CCTTGAGCTCTTAGTGGCGCAGCTAACGCATTAAGTTGACCGCCTGGGGAGTACGGCCGCAAGGTTAAAACTCAAATGAA",
      "TTGACGG"
    )
  )
  blastASVs(seqs1, seqs2, cutoff = 0)
}

}
\keyword{BLAST}
