% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeFilter.R
\name{edgeFilter}
\alias{edgeFilter}
\title{Function to filter edges from networks generated by \code{asvNet}.}
\usage{
edgeFilter(net, filter, edge = "spearman_distance")
}
\arguments{
\item{net}{Object returned from \link{asvNet}.}

\item{filter}{Value to filter edges for. If non-NULL then only edges with edgeWeight greater than
this value are kept. This can be a character vector or a numeric, with character taken as a quantile
and numeric taken as a value to use as is.}

\item{edge}{Optional weighting for edges. Must be present in the "edges" of net. Default of NULL
will show equal size edges between all connected nodes.}
}
\value{
A modified version of net with filtered edges (and nodes if any were now isolated).
}
\description{
Function to filter edges from networks generated by \code{asvNet}.
}
\examples{

taxa <- c(
  "Bacteria", "Proteobacteria", "Betaproteobacteria", "Burkholderiales",
  "Burkholderiaceae", "Paraburkholderia", NA
)
taxa <- matrix(rep(taxa, 10), nrow = 10, byrow = TRUE)
colnames(taxa) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
rownames(taxa) <- paste0("ASV", 1:10)
# taxonomy data if used should have ASV names explicitly as a column
taxa_df <- as.data.frame(taxa)
taxa_df$asv <- rownames(taxa_df)

sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)
net_data <- asvNet(sp_dist, taxa_df, edge = "spearman_distance")
dim(net_data$edges)
net_data2 <- edgeFilter(net_data, 0.6)
dim(net_data2$edges)

}
