% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_class.R
\name{intersect_thresh}
\alias{intersect_thresh}
\title{Getting intersections of thresh objects}
\usage{
intersect_thresh(x, y)
}
\arguments{
\item{x}{thresh object}

\item{y}{thresh object}
}
\description{
Getting intersections of thresh objects
}
