% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netPlot.R
\name{net.plot}
\alias{net.plot}
\title{Function to plot networks generated from \code{asvNet} with various emphases.}
\usage{
net.plot(
  net,
  fill = NULL,
  shape = NULL,
  size = 3,
  edgeWeight = NULL,
  edgeFilter = NULL,
  thresh_below = 0.05
)
}
\arguments{
\item{net}{Object returned from \code{\link{asvNet}}}

\item{fill}{Optional column name to fill points by. Accepts single column names, vectors of p-value columns, or "thresh" which will match all p-value columns if the network was fit with threshold model data from \link{\code{thresh}}.}

\item{shape}{Optional column name to use for node shapes. Accepts the same options as fill.}

\item{size}{Size for points. Passed to ggplot2::geom_point.}

\item{edgeWeight}{Optional weighting for edges. Must be present in the "edges" of net. Default of NULL will show equal size edges between all connected nodes.}

\item{edgeFilter}{Optional value to filter edges for. If non-NULL then only edges with edgeWeight greater than this value are kept. This can be a character vector or a numeric. Character vectors are interpreted as quantiles ("0.5" corresponds to the top 50% are kept).}

\item{thresh_below}{Significant cutoff if p-value columns are used for fill or shape. Defaults to 0.05.}
}
\value{
A named list with three elements:
"Nodes" is a dataframe of nodes and their metadata
"Edges" is a dataframe of edges connecting nodes.
"graph" is the igraph object used to generate the dataframes.
}
\description{
Function to plot networks generated from \code{asvNet} with various emphases.
}
\examples{

# a<-qc(); b<-cal(a); c<-thresh(b); d<-asvDist(a) ; e<-net(d, thresh = c)
print(load("/home/jsumner/Desktop/stargate/SINC/sincUtils/syncomBuilder/net_output.rdata"))

net = net_data
fill=NULL
shape=NULL
size=3
edgeWeight="spearman"
edgeFilter=NULL
thresh_below = 0.05
net.plot(net, fill, shape, size, edgeWeight, edgeFilter, thresh_below)
net.plot(net, fill="thresh", shape, size, edgeWeight, edgeFilter, thresh_below)


}
\keyword{changepoint}
\keyword{network,}
