% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSamples.R
\name{rankSamples}
\alias{rankSamples}
\title{Function to pick samples for limited dilution/other bench science following changepoint
regression of an ASV table or network.}
\usage{
rankSamples(
  asvTab,
  thresh,
  network = NULL,
  id = "sample",
  groups = NULL,
  phenotypes = NULL,
  p.cutoff = 0.05
)
}
\arguments{
\item{asvTab}{An asv table as returned by \code{\link{qc}} or \code{\link{cal}}.}

\item{thresh}{Output from \code{\link{thresh}} or \code{\link{netThresh}}.}

\item{network}{Optionally a network object as returned from \code{\link{asvNet}} and clustered
with \code{\link{netClust}}. Defaults to NULL in which case the asv table is used and thresh
is assumed to be changepoint regressions at the asv level.}

\item{id}{The name of the column containing sample names, defaults to "sample".}

\item{groups}{Optionally specify which ASVs/Clusters should be used. If NULL, the default,
then significant models (p value below p.cutoff) are used to find groups.}

\item{phenotypes}{Optionally specify which phenotypes should be used. If NULL, the default,
then all in thresh are used.}

\item{p.cutoff}{P value cutoff for statistical significance if groups is NULL. Defaults to 0.05.}
}
\value{
A data frame of samples ranked by the impact of their nodes (ASVs) or clusters
}
\description{
Function to pick samples for limited dilution/other bench science following changepoint
regression of an ASV table or network.
}
\examples{

asv$biomass_z <- rnorm(nrow(asv))
tm <- thresh(asv, "biomass_z")
x <- rankSamples(asv, tm, id = "sample")

}
