% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSamples.R
\name{rankSamples}
\alias{rankSamples}
\alias{rankSamples.thresh}
\alias{rankSamples.data.frame}
\title{Function to pick samples for limited dilution/other bench science following changepoint
regression of an ASV table or network.}
\usage{
rankSamples(x, ...)

\method{rankSamples}{thresh}(x, cutoff = 0.05, asvTab = NULL, ...)

\method{rankSamples}{data.frame}(x, cols = NULL, ...)
}
\arguments{
\item{x}{ASV table as a data.frame or a \code{thresh} object. For a data.frame the cols argument
must be specified. For a \code{thresh} object all ASVs that were significantly related to a
phenotype are used (significance determined by the cutoff argument).}

\item{...}{Additional arguments passed to methods.}

\item{cutoff}{P value cutoff for a thresh model being significant}

\item{asvTab}{Optional asv table. If missing then the data slot of the thresh object is used.}

\item{cols}{Columns to use in scoring. These should be some subset of ASV columns.}
}
\value{
A data frame of samples ranked by the impact of their nodes (ASVs) or clusters
}
\description{
Function to pick samples for limited dilution/other bench science following changepoint
regression of an ASV table or network.
}
\examples{

set.seed(123)
asv$biomass_z <- rnorm(nrow(asv))
x <- thresh(asv, "biomass_z")
rankSamples(x, cutoff = 0.5)

}
