% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaAg.R
\name{taxaAg}
\alias{taxaAg}
\title{Aggregate an ASV table to some taxonomic level}
\usage{
taxaAg(x = NULL, tx = NULL, taxaLevel = "Class")
}
\arguments{
\item{x}{ASV table, asv columns should start with "ASV" prefix.}

\item{tx}{Taxonomy table in dada2 format. Rownames should correspond to ASV names in the same format
as x's column names. If this is a matrix it will be converted to a data.frame.}

\item{taxaLevel}{A level of the taxonomy in `tx` to aggregate data to. Must match a colname of `tx`.}
}
\value{
A list of two dataframes. "data" is similar to an ASV table but aggregated to the
given taxaLevel.
"meta" is a summary of the aggregation containing the number of ASVs and the total number of counts
in each family in long format.
}
\description{
This is broadly useful for visualizing taxonomic compositions and coarse graining for networks,
ordinations, or other analysis where the ASV level of granularity is either too specific or
confounded by some effect (biological or computational).
}
\examples{
x <- taxaAg(asv, taxa, "Class")
lapply(x, head)

}
\keyword{dirichlet}
\keyword{network}
