% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tresh_class.R
\docType{class}
\name{thresh-class}
\alias{thresh-class}
\title{Class \code{thresh} for output from \code{syncomBuildR::thresh} and related functions.}
\description{
Class for changepoint models as used in syncomBuildR.
}
\details{
See \code{methods(class = "thresh")} for an overview of available methods.
}
\section{Slots}{

\describe{
\item{\code{intercept}}{The intercept term (y value pre changepoint, per hinge model)}

\item{\code{changepoint}}{The changepoint term (x value at which slope takes effect)}

\item{\code{slope}}{The slope term}

\item{\code{phenotype}}{The phenotype (outcome) that the model was fit to.}

\item{\code{model}}{the model object}

\item{\code{predictor}}{the ASV/Cluster (predictor) that the model was fit using.}

\item{\code{data}}{The data used to fit the model}

\item{\code{type}}{The backend used to fit the model as well as the type of model. Currently "chngptm" is
supported.}

\item{\code{unit}}{The unit/scope. Current options are "individual" for single ASVs or "cluster" for groups
of ASVs.}

\item{\code{control}}{Additional information such as the call.}
}}

\seealso{
\code{\link{thresh}}
}
