% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.R
\name{thresh}
\alias{thresh}
\alias{thresh.scbnet}
\title{Function to run changepoint models of individual microbes vs phenotypes or
communities/clusters vs phenotypes.}
\usage{
thresh(
  x,
  phenoCols,
  predCols = NULL,
  model = "hinge",
  cores = getOption("mc.cores", 1),
  calibratePheno = NULL,
  p.adjust.method = "none",
  ...
)

\method{thresh}{scbnet}(
  x,
  phenoCols,
  predCols = NULL,
  model = "hinge",
  cores = getOption("mc.cores", 1),
  calibratePheno = NULL,
  p.adjust.method = "none",
  asvTab = NULL,
  ...
)
}
\arguments{
\item{x}{An asv table or analogous dataframe with a row per observation and columns for traits.}

\item{phenoCols}{A character vector of column names for phenotypes to be used in changepoint models.}

\item{predCols}{A character vector of columns representing microbes (predictor variables).
Defaults NULL where all column names containing the string "ASV" will be used.}

\item{model}{Type of changepoint model in chngpt::chngptm labeling convention. Currently hinge,
upperhinge, and segmented are supported. See Figure 2.1 of the chngpt
[vignette](https://cran.r-project.org/web/packages/chngpt/vignettes/chngpt-vignette.pdf).}

\item{cores}{Number of cores to run in parallel.}

\item{calibratePheno}{An optional vector of column names to calibrate the phenotypes by.
This should generally correspond to those used in `cal` if the ASV table has been calibrated or
just represent confounders that you wish to remove effects from in the changepoint regression.
\code{lm} is used to get residuals of the phenotype after these effects are regressed out.}

\item{p.adjust.method}{A method to adjust for multiple testing, defaults to "none". Available options
are shown with \code{stats::p.adjust.methods}.}

\item{...}{Additional arguments passed to methods.}

\item{asvTab}{An ASV table used in making the \code{scbnet} object.}
}
\value{
A \code{thresh} object.
}
\description{
Function to run changepoint models of individual microbes vs phenotypes or
communities/clusters vs phenotypes.
}
\examples{

asv$biomass_z <- rnorm(nrow(asv))
tm <- thresh(asv, "biomass_z")
tm

taxa <- c(
  "Bacteria", "Proteobacteria", "Betaproteobacteria", "Burkholderiales",
  "Burkholderiaceae", "Paraburkholderia", NA
)
taxa <- matrix(rep(taxa, 10), nrow = 10, byrow = TRUE)
colnames(taxa) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
rownames(taxa) <- paste0("ASV", 1:10)
# taxonomy data if used should have ASV names explicitly as a column
taxa_df <- as.data.frame(taxa)
taxa_df$asv <- rownames(taxa_df)
sp_dist <- asvDist(asv, method = "spearman", clr_transform = TRUE, edgeFilter = 0.5)
net_data <- asvNet(sp_dist, taxa_df, edge = "spearman_similarity")
net_data <- netClust(net = net_data, "components")
net_data <- netClust(net = net_data, pullNode(net_data,
  node = "ASV10",
  edge = "spearman_similarity", edgeFilter = 0.7, nodeCol = "asv"
))
net_data <- netClust(net = net_data, pullNode(net_data,
  node = "ASV1",
  edge = "spearman_similarity", edgeFilter = 0.7, nodeCol = "asv"
))
asv$biomass_z <- rnorm(nrow(asv))
tn <- thresh(net_data, phenoCols = "biomass_z", asvTab = asv)


# ARGS
x <- net_data
phenoCols = "biomass_z"
predCols = NULL
model = "hinge"
cores = getOption("mc.cores", 1)
calibratePheno = NULL
p.adjust.method = "none"
asvTab = asv
}
\keyword{changepoint,}
\keyword{phenotype}
\keyword{regression,}
\keyword{threshold,}
