% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshUpset.R
\name{threshUpset}
\alias{threshUpset}
\title{Function to plot results of a changepoint models as upset plots}
\usage{
threshUpset(thresh, cutoff = 0.05)
}
\arguments{
\item{thresh}{Output from \code{\link{thresh}}}

\item{cutoff}{Alpha, the significance threshold. Defaults to 0.05.}
}
\value{
A ggplot or list of ggplots showing changepoint models against some set of phenotypes.
}
\description{
Function to plot results of a changepoint models as upset plots
}
\examples{

set.seed(123)
asv$biomass_z <- rnorm(nrow(asv))
asv$height_z <- rnorm(nrow(asv))
tm <- thresh(asv, c("biomass_z", "height_z"))

threshUpset(tm, cutoff = 0.5)

}
\keyword{changepoint,}
\keyword{phenotype,}
\keyword{regression,}
\keyword{threshold,}
\keyword{upset}
