% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshUpset.R
\name{threshUpset}
\alias{threshUpset}
\title{Function to plot results of a changepoint models as upset plots}
\usage{
threshUpset(
  thresh,
  mode = "phenotype",
  pvalcol = "p.value.",
  cutoff = 0.05,
  cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{thresh}{Output from \code{\link{thresh}}}

\item{mode}{Either "phenotype" (the default) or "pvalue". "phenotype" will show an upset plot
of significant changepoint models against the phenotypes present in thresh, "pvalue" will show
the results of various p-value adjustment methods.}

\item{pvalcol}{The column in thresh that has p-value data, defaults to "p.value." which reflects
standard output from \code{thresh}.}

\item{cutoff}{Alpha, the significance threshold. Defaults to 0.05.}

\item{cores}{Number of cores to use in parallel. This should not make much of a difference generally
but can be useful with very large tables. Defaults to 1 if the "mc.cores" option is not set.}
}
\value{
A ggplot or list of ggplots showing changepoint models against some set of phenotypes.
}
\description{
Function to plot results of a changepoint models as upset plots
}
\examples{

# a<-qc(); b<-cal(a); c<-thresh(b)
print(load("~/scripts/SINC/sincUtils/syncomBuilder/threshOutput.rdata"))
head(threshMods)
threshMods[1:317, "phenotype"] <- "not_biomass"
threshUpset(threshMods)

}
\keyword{changepoint,}
\keyword{phenotype,}
\keyword{regression,}
\keyword{threshold,}
\keyword{upset}
