% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapdf.R
\name{umapdf}
\alias{umapdf}
\title{Function to run a UMAP, plot and return the data with coordinates}
\usage{
umapdf(df = NULL, cols = NULL, color = NULL, returnData = TRUE, ...)
}
\arguments{
\item{df}{Dataframe to ordinate. Generally UMAP works best on principal components,
so this is assumed to be output from \link{pcadf}.}

\item{cols}{columns to reduce dimensions of. Can be specified with names or positions.
Defaults to all column names starting with "pc" then numbers for use with \link{pcadf}.}

\item{color}{column name used to color points in the umap plot.}

\item{returnData}{Logical, should data be returned?
Defaults to TRUE where data and a ggplot are returned.}

\item{...}{Additional arguments passed to \code{uwot::umap}. The n_neighbors, n_components,
and pca arguments are potentially commonly useful.}
}
\description{
Function to run a UMAP, plot and return the data with coordinates
}
\examples{
print(load("~/scripts/SINC/sincUtils/syncomBuilder/cal_output.rdata"))
asv <- are_c[[1]]
zinbCalibrated <- are_c[[2]][are_c[[2]]$model == "ZINB", "asv"]
asv <- are_c[[1]][, c("tissue", "plot", "row", "genotype", "biomass", "sd", zinbCalibrated)]
pdf <- pcadf(df = asv, cols = NULL, color = c("tissue", "genotype"), returnData = TRUE, ncp = NULL)
udf <- umapdf(df = pdf$data, cols = NULL, color = c("tissue", "genotype"),
returnData = TRUE, pca = 50)

}
\keyword{pca,}
\keyword{umap}
